Several Changes made: Databases: I added a new method called clearData(). This clears the data from the database. This method should be called after the database has been instantiated inside test classes.
HistoryActivityTest: The Database that was being used was the Drafts Database. It should have been the Message Database. I changed it. Second, the second test was incorrect. In order to test if the method is functional, the object messagesToBeDisplayed must be checked with an assertion to see if it has the expected amount of messages that are to be displayed. I changed the test so that it will work.